- added filesystem check for unified write extend. If first   mount failed, try to check filesystem, if next mount failed   re-create the write filesystem of the unified tree - added support for unified virtual disk images to oemboot.   In this case the kiwi initrd will not be replaced by the   distribution default initrd
git-svn-id: https://svn.berlios.de/svnroot/repos/kiwi/kiwi-head@609 92bc7f66-5d17-0410-ac0d-e081219a1bb3

+++ modules/KIWILinuxRC.sh
@@ -1044,14 +1044,18 @@ function mountSystem () {
 					test "$RELOAD_IMAGE" = "yes" || \
 					! mount $rwDevice $rwDir >/dev/null 2>&1
 				then
-					Echo "Creating EXT2 filesystem for RW data on $rwDevice..."
-					if ! mke2fs $rwDevice >/dev/null 2>&1;then
-						systemException \
-							"Failed to create ext2 filesystem" \
-						"reboot"
-					fi
-					Echo "Checking EXT2 write extend..."
+					Echo "Checking filesystem for RW data on $rwDevice..."
 					e2fsck -y -f $rwDevice >/dev/null 2>&1
+					if ! mount $rwDevice $rwDir >/dev/null 2>&1;then
+						Echo "Creating filesystem for RW data on $rwDevice..."
+						if ! mke2fs $rwDevice >/dev/null 2>&1;then
+							systemException \
+								"Failed to create ext2 filesystem" \
+							"reboot"
+						fi
+						Echo "Checking EXT2 write extend..."
+						e2fsck -y -f $rwDevice >/dev/null 2>&1
+					fi
 				else
 					umount $rwDevice
 				fi
